backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2
python-dotenv
retell-sdk
redis
twilio
pyjwt[crypto]
psycopg2-binary
asyncpg
httpx
cryptography